requirements.txt
flask
wtforms
requests
Pillow
pycairo
svgwrite
jieba
gensim